+++ apps/mobile/src/app/account/[accountId]/activity.tsx
@@ -1,6 +1,9 @@
 import { configureAccountParamsSchema } from '@/app/account/[accountId]/index';
-import { FetchWrapper } from '@/components/loading';
-import { ActivityEmpty, ActivityLayout, ActivityList, hasActivity } from '@/features/activity';
+import { FetchWrapper } from '@/components/loading/fetch-wrapper';
+import { ActivityEmpty } from '@/features/activity/activity-empty';
+import { ActivityList } from '@/features/activity/activity-list';
+import { hasActivity } from '@/features/activity/activity-widget';
+import { ActivityLayout } from '@/features/activity/activity.layout';
 import { useAccountActivity } from '@/queries/activity/account-activity.query';
 import { deserializeAccountId } from '@/store/accounts/accounts';
 import { useLocalSearchParams } from 'expo-router';
