+++ apps/mobile/src/app/account/[accountId]/collectibles.tsx
@@ -1,5 +1,5 @@
 import { AnimatedHeaderScreenLayout } from '@/components/headers/animated-header/animated-header-screen.layout';
-import { Collectibles } from '@/features/collectibles';
+import { Collectibles } from '@/features/collectibles/collectibles';
 import { NetworkBadge } from '@/features/settings/network-badge';
 import { useAccountCollectibles } from '@/queries/collectibles/account-collectibles.query';
 import { deserializeAccountId } from '@/store/accounts/accounts';
