+++ apps/mobile/src/app/collectibles/index.tsx
@@ -1,6 +1,6 @@
 import { AnimatedHeaderScreenLayout } from '@/components/headers/animated-header/animated-header-screen.layout';
 import { PageLayout } from '@/components/page/page.layout';
-import { Collectibles } from '@/features/collectibles';
+import { Collectibles } from '@/features/collectibles/collectibles';
 import { NetworkBadge } from '@/features/settings/network-badge';
 import { useTotalCollectibles } from '@/queries/collectibles/account-collectibles.query';
 import { t } from '@lingui/macro';
