+++ apps/mobile/src/app/index.tsx
@@ -6,7 +6,8 @@ import { ActivityWidget } from '@/features/activity/activity-widget';
 import { AllAccountBalances } from '@/features/balances/balances';
 import { BalancesWidget } from '@/features/balances/balances-widget';
 import { TotalBalance } from '@/features/balances/total-balance';
-import { Collectibles, CollectiblesWidget, hasCollectibles } from '@/features/collectibles';
+import { Collectibles, hasCollectibles } from '@/features/collectibles/collectibles';
+import { CollectiblesWidget } from '@/features/collectibles/collectibles-widget';
 import { EarnWidget } from '@/features/earn/earn-widget';
 import { useCollectiblesFlag } from '@/features/feature-flags';
 import { NotificationsSheet } from '@/features/notifications/notifications-sheet';
