--- apps/mobile/src/components/loading/index.ts
@@ -1,5 +1,0 @@
-export * from './empty-layout';
-export * from '../error/fetch-error';
-export * from './fetch-state';
-export * from './fetch-wrapper';
-export * from './loading';
