--- apps/mobile/src/components/widget/index.ts
@@ -1,3 +1,0 @@
-export * from './widget';
-export * from './widget-header';
-export * from './widget-title';
