--- apps/mobile/src/features/activity/index.ts
@@ -1,6 +1,0 @@
-export * from './activity-list-item';
-export * from './activity-empty';
-export * from './activity-widget';
-export * from './activity.layout';
-export * from './activity-list';
-export * from './activity-list-item';
