+++ apps/mobile/src/features/balances/bitcoin/runes-balance.tsx
@@ -1,4 +1,5 @@
-import { FetchState, FetchWrapper } from '@/components/loading';
+import { FetchState } from '@/components/loading/fetch-state';
+import { FetchWrapper } from '@/components/loading/fetch-wrapper';
 import { BalanceViewProps } from '@/features/balances/balances';
 import {
   useRunesAccountBalance,
