--- apps/mobile/src/features/collectibles/index.ts
@@ -1,4 +1,0 @@
-export * from './collectibles';
-export * from './collectibles.layout';
-export * from './collectibles-serializer';
-export * from './collectibles-widget';
