--- apps/mobile/src/features/wallet-manager/add-wallet/index.ts
@@ -1,1 +1,0 @@
-export { AddWalletSheet } from './add-wallet-sheet';
